+++ app/src/endpoints.c
@@ -130,7 +130,7 @@ static int send_consumer_report() {
     }
 }
 
-int zmk_endpoints_send_report(uint8_t usage_page) {
+int zmk_endpoints_send_report(uint16_t usage_page) {
 
     LOG_DBG("usage page 0x%02X", usage_page);
     switch (usage_page) {
